+++ src/Urls/MemoizedUrlRecord.php
@@ -6,11 +6,40 @@ use Illuminate\Database\Eloquent\Model;
 
 class MemoizedUrlRecord extends UrlRecord
 {
+    static $cachedRecords;
+
+    public static function clearCachedRecords()
+    {
+        static::$cachedRecords = null;
+    }
+
+    /**
+     * Here we cache all the url records and determine the proper url record
+     * via the collection methods. This is a lot faster on large data sets.
+     *
+     * @param Model $model
+     * @param string|null $locale
+     * @return UrlRecord
+     * @throws UrlRecordNotFound
+     */
     public static function findByModel(Model $model, string $locale = null): UrlRecord
     {
-        return chiefMemoize('url-records-find-by-model', function ($model, $locale = null) {
-            return parent::findByModel($model, $locale);
-        }, [$model, $locale]);
+        if(!static::$cachedRecords) {
+            static::$cachedRecords = parent::all();
+        }
+
+        $record = static::$cachedRecords
+            ->where('model_type', $model->getMorphClass())
+            ->where('model_id', $model->id)
+            ->where('locale', $locale)
+            ->sortBy('redirect_id')
+            ->first();
+
+        if (!$record) {
+            throw new UrlRecordNotFound('No url record found for model ['.$model->getMorphClass().'@'.$model->id.'] for locale ['.$locale.'].');
+        }
+
+        return $record;
     }
 
     public static function getByModel(Model $model)