+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All Notable changes to the `chief` application template will be documented in th
 ## unreleased
 - Fixed: fixed bug with set viewkey
 - Fixed: required on settings
+- Changed: optimize url record retrieval for large datasets with a simple memoization.
 - Changed: small changes to pagebuilder and removed redactor
 
 ## 0.3.0 - 2019-8-20